requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.9
pydantic>=2.7.0
python-dotenv>=1.0.1
crewai>=0.80.0
crewai-tools>=0.14.0
httpx>=0.27.0
